pcap: insert VLAN accel data into the payload
When using AF_PACKET utilities, if the VLAN information is hw accelerated it is
inserted in the kernel before the packet reaches userspace. Because of this,
those tools display VLAN data as part of the packet even when accelerated.

This is not the case in Retis, as we report the packet as-is and the VLAN hw
acceleration data as part of the event. When using Retis, both cases work well
as the VLAN data is printed one way or another (in the packet or as an event
section). But when converting the packet to the PCAP format, the result is hw
acceleration data is not inserted and the result differs from what we can
expected when using AF_PACKET utilities.

Insert the VLAN tag when accelerated by default when converting events to PCAP
data in Retis, to match the other utilities expectations. This can be disabled
with a cli flag.

Signed-off-by: Antoine Tenart <atenart@redhat.com>