events: do not implement Copy for event types
Clone is as efficient as Copy for types implementing Copy. Not implementing Copy
on event types embedding Copy-only fields should not change performances.

This is required for later changes, but also makes sense as event types can be
large and silent copying them can be an issue. This changes helped highlighting
some problematic uses (fixed alongside those changes). The drawback is a few
places are a bit more verbose now, especially when dealing with TimeSpec.

Signed-off-by: Antoine Tenart <atenart@redhat.com>